AI/API/hand/requirements.txt
Flask
flask-cors
Pillow
opencv-python
numpy
pytesseract
